Refactor Tezos formats a bit
